status changes in db

+++ database/migrations/2024_03_30_060114_add_status_to_applications_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('applications', function (Blueprint $table) {
-            $table->tinyInteger('status')->default(0)->comment('0: Initiated, 1: In Progress, 2: Completed');
+            $table->unsignedBigInteger('status')->default(0)->comment('0: Initiated, 1: In Progress, 2: Completed');
         });
     }
